Keep only fileattrib, createdts & GenericAttrib
