+++ doc/050_restore.rst
@@ -72,10 +72,6 @@ Restoring symbolic links on windows is only possible when the user has
 ``SeCreateSymbolicLinkPrivilege`` privilege or is running as admin. This is a
 restriction of windows not restic.
 
-Restoring security descriptors on windows is only possible when the user has
-``SeBackupPrivilege``, ``SeRestorePrivilege`` and ``SeSecurityPrivilege`` 
-privilege or is running as admin. This is a restriction of windows not restic.
-
 By default, restic does not restore files as sparse. Use ``restore --sparse`` to
 enable the creation of sparse files if supported by the filesystem. Then restic
 will restore long runs of zero bytes as holes in the corresponding files.
