+++ internal/restic/node_windows_test.go
@@ -4,284 +4,17 @@
 package restic_test
 
 import (
-	"encoding/base64"
 	"encoding/binary"
 	"fmt"
 	"os"
 	"path/filepath"
-	"strings"
 	"syscall"
 	"testing"
 
 	"github.com/restic/restic/internal/errors"
-	"github.com/restic/restic/internal/fs"
 	"github.com/restic/restic/internal/restic"
 	"github.com/restic/restic/internal/test"
-	"golang.org/x/sys/windows"
 )
-
-func TestRestoreExtendedAttributes(t *testing.T) {
-	tempDir := t.TempDir()
-	expectedNodes := []restic.Node{
-		{
-			Name:       "testfile",
-			Type:       "file",
-			Mode:       0644,
-			ModTime:    parseTime("2005-05-14 21:07:03.111"),
-			AccessTime: parseTime("2005-05-14 21:07:04.222"),
-			ChangeTime: parseTime("2005-05-14 21:07:05.333"),
-			ExtendedAttributes: []restic.ExtendedAttribute{
-				{"user.foo", []byte("bar")},
-			},
-		},
-		{
-			Name:       "testdirectory",
-			Type:       "dir",
-			Mode:       0755,
-			ModTime:    parseTime("2005-05-14 21:07:03.111"),
-			AccessTime: parseTime("2005-05-14 21:07:04.222"),
-			ChangeTime: parseTime("2005-05-14 21:07:05.333"),
-			ExtendedAttributes: []restic.ExtendedAttribute{
-				{"user.foo", []byte("bar")},
-			},
-		},
-	}
-	for _, testNode := range expectedNodes {
-		testPath := filepath.Join(tempDir, "001", testNode.Name)
-		if err := os.MkdirAll(filepath.Dir(testPath), testNode.Mode); err != nil {
-			t.Fatalf("Failed to create parent directories: %v", err)
-		}
-		if testNode.Type == "file" {
-
-			testFile, err := os.Create(testPath)
-			if err != nil {
-				t.Fatalf("Failed to create test file: %v", err)
-			}
-			testFile.Close()
-		} else if testNode.Type == "dir" {
-
-			err := os.Mkdir(testPath, testNode.Mode)
-			if err != nil {
-				t.Fatalf("Failed to create test directory: %v", err)
-			}
-		}
-
-		err := testNode.RestoreMetadata(testPath)
-		if err != nil {
-			t.Fatalf("Error restoring metadata: %v", err)
-		}
-		var handle windows.Handle
-		utf16Path := windows.StringToUTF16Ptr(testPath)
-		if testNode.Type == "file" {
-			handle, err = windows.CreateFile(utf16Path, windows.FILE_READ_EA, 0, nil, windows.OPEN_EXISTING, windows.FILE_ATTRIBUTE_NORMAL, 0)
-		} else if testNode.Type == "dir" {
-			handle, err = windows.CreateFile(utf16Path, windows.FILE_READ_EA, 0, nil, windows.OPEN_EXISTING, windows.FILE_ATTRIBUTE_NORMAL|windows.FILE_FLAG_BACKUP_SEMANTICS, 0)
-		}
-		if err != nil {
-			t.Fatalf("Error opening file/directory: %v", err)
-		}
-		defer func() {
-			err := windows.Close(handle)
-			if err != nil {
-				t.Logf("Error closing file %s: %v\n", testPath, err)
-			}
-		}()
-
-		if len(testNode.ExtendedAttributes) > 0 {
-			extAttr, err := fs.GetFileEA(handle)
-			if err != nil {
-				t.Fatalf("Error getting extended attributes: %v", err)
-			}
-			test.Equals(t, len(testNode.ExtendedAttributes), len(extAttr))
-
-			for _, expectedExtAttr := range testNode.ExtendedAttributes {
-				var foundExtAttr *fs.ExtendedAttribute
-				for _, ea := range extAttr {
-					if strings.EqualFold(ea.Name, expectedExtAttr.Name) {
-						foundExtAttr = &ea
-						break
-
-					}
-				}
-				test.Assert(t, foundExtAttr != nil, "Expected extended attribute not found")
-				test.Equals(t, expectedExtAttr.Value, foundExtAttr.Value)
-			}
-		}
-	}
-}
-
-func TestRestoreSecurityDescriptors(t *testing.T) {
-	tempDir := t.TempDir()
-	expectedNodes := []restic.Node{
-		{
-			Name:       "testfile",
-			Type:       "file",
-			Mode:       0644,
-			ModTime:    parseTime("2005-05-14 21:07:03.111"),
-			AccessTime: parseTime("2005-05-14 21:07:04.222"),
-			ChangeTime: parseTime("2005-05-14 21:07:05.333"),
-			ExtendedAttributes: []restic.ExtendedAttribute{
-				{"user.foo", []byte("bar")},
-			},
-			GenericAttributes: []restic.GenericAttribute{
-				restic.NewGenericAttribute(restic.TypeSecurityDescriptor, []byte("AQAUrBQAAAAwAAAAAAAAAEwAAAABBQAAAAAABRUAAACIn1iuVqCC6sy9JqvqAwAAAQUAAAAAAAUVAAAAiJ9YrlaggurMvSarAQIAAAIAfAAEAAAAAAAkAKkAEgABBQAAAAAABRUAAACIn1iuVqCC6sy9JqvtAwAAABAUAP8BHwABAQAAAAAABRIAAAAAEBgA/wEfAAECAAAAAAAFIAAAACACAAAAECQA/wEfAAEFAAAAAAAFFQAAAIifWK5WoILqzL0mq+oDAAA=")),
-			},
-		},
-		{
-			Name:       "testfile2",
-			Type:       "file",
-			Mode:       0644,
-			ModTime:    parseTime("2005-05-14 21:07:03.111"),
-			AccessTime: parseTime("2005-05-14 21:07:04.222"),
-			ChangeTime: parseTime("2005-05-14 21:07:05.333"),
-			ExtendedAttributes: []restic.ExtendedAttribute{
-				{"user.foo2", []byte("bar2")},
-			},
-			GenericAttributes: []restic.GenericAttribute{
-				restic.NewGenericAttribute(restic.TypeSecurityDescriptor, []byte("AQAUjBQAAAAwAAAA7AAAAEwAAAABBQAAAAAABRUAAAAvr7t03PyHGk2FokNHCAAAAQUAAAAAAAUVAAAAiJ9YrlaggurMvSarAQIAAAIAoAAFAAAAAAAkAP8BHwABBQAAAAAABRUAAAAvr7t03PyHGk2FokNHCAAAAAAkAKkAEgABBQAAAAAABRUAAACIn1iuVqCC6sy9JqvtAwAAABAUAP8BHwABAQAAAAAABRIAAAAAEBgA/wEfAAECAAAAAAAFIAAAACACAAAAECQA/wEfAAEFAAAAAAAFFQAAAIifWK5WoILqzL0mq+oDAAACAHQAAwAAAAKAJAC/AQIAAQUAAAAAAAUVAAAAL6+7dNz8hxpNhaJDtgQAAALAJAC/AQMAAQUAAAAAAAUVAAAAL6+7dNz8hxpNhaJDPgkAAAJAJAD/AQ8AAQUAAAAAAAUVAAAAL6+7dNz8hxpNhaJDtQQAAA==")),
-			},
-		},
-		{
-			Name:       "testdirectory",
-			Type:       "dir",
-			Mode:       0755,
-			ModTime:    parseTime("2005-05-14 21:07:03.111"),
-			AccessTime: parseTime("2005-05-14 21:07:04.222"),
-			ChangeTime: parseTime("2005-05-14 21:07:05.333"),
-			ExtendedAttributes: []restic.ExtendedAttribute{
-				{"user.foo3", []byte("bar3")},
-			},
-			GenericAttributes: []restic.GenericAttribute{
-				restic.NewGenericAttribute(restic.TypeSecurityDescriptor, []byte("AQAUrBQAAAAwAAAAAAAAAEwAAAABBQAAAAAABRUAAACIn1iuVqCC6sy9JqvqAwAAAQUAAAAAAAUVAAAAiJ9YrlaggurMvSarAQIAAAIAfAAEAAAAAAAkAKkAEgABBQAAAAAABRUAAACIn1iuVqCC6sy9JqvtAwAAABMUAP8BHwABAQAAAAAABRIAAAAAExgA/wEfAAECAAAAAAAFIAAAACACAAAAEyQA/wEfAAEFAAAAAAAFFQAAAIifWK5WoILqzL0mq+oDAAA=")),
-			},
-		},
-		{
-			Name:       "testdirectory2",
-			Type:       "dir",
-			Mode:       0755,
-			ModTime:    parseTime("2005-05-14 21:07:03.111"),
-			AccessTime: parseTime("2005-05-14 21:07:04.222"),
-			ChangeTime: parseTime("2005-05-14 21:07:05.333"),
-			ExtendedAttributes: []restic.ExtendedAttribute{
-				{"user.foo4", []byte("bar4")},
-			},
-			GenericAttributes: []restic.GenericAttribute{
-				restic.NewGenericAttribute(restic.TypeSecurityDescriptor, []byte("AQAUrBQAAAAwAAAAAAAAAEwAAAABBQAAAAAABRUAAACIn1iuVqCC6sy9JqvqAwAAAQUAAAAAAAUVAAAAiJ9YrlaggurMvSarAQIAAAIA3AAIAAAAAAIUAKkAEgABAQAAAAAABQcAAAAAAxQAiQASAAEBAAAAAAAFBwAAAAAAJACpABIAAQUAAAAAAAUVAAAAiJ9YrlaggurMvSar7QMAAAAAJAC/ARMAAQUAAAAAAAUVAAAAiJ9YrlaggurMvSar6gMAAAALFAC/ARMAAQEAAAAAAAMAAAAAABMUAP8BHwABAQAAAAAABRIAAAAAExgA/wEfAAECAAAAAAAFIAAAACACAAAAEyQA/wEfAAEFAAAAAAAFFQAAAIifWK5WoILqzL0mq+oDAAA=")),
-			},
-		},
-	}
-	for _, testNode := range expectedNodes {
-		testPath := filepath.Join(tempDir, "001", testNode.Name)
-		if err := os.MkdirAll(filepath.Dir(testPath), testNode.Mode); err != nil {
-			t.Fatalf("Failed to create parent directories: %v", err)
-		}
-		if testNode.Type == "file" {
-
-			testFile, err := os.Create(testPath)
-			if err != nil {
-				t.Fatalf("Failed to create test file: %v", err)
-			}
-			testFile.Close()
-		} else if testNode.Type == "dir" {
-
-			err := os.Mkdir(testPath, testNode.Mode)
-			if err != nil {
-				t.Fatalf("Failed to create test directory: %v", err)
-			}
-		}
-
-		err := testNode.RestoreMetadata(testPath)
-		if err != nil {
-			t.Fatalf("Error restoring metadata: %v", err)
-		}
-		var handle windows.Handle
-		utf16Path := windows.StringToUTF16Ptr(testPath)
-		if testNode.Type == "file" {
-			handle, err = windows.CreateFile(utf16Path, windows.FILE_READ_EA, 0, nil, windows.OPEN_EXISTING, windows.FILE_ATTRIBUTE_NORMAL, 0)
-		} else if testNode.Type == "dir" {
-			handle, err = windows.CreateFile(utf16Path, windows.FILE_READ_EA, 0, nil, windows.OPEN_EXISTING, windows.FILE_ATTRIBUTE_NORMAL|windows.FILE_FLAG_BACKUP_SEMANTICS, 0)
-		}
-		if err != nil {
-			t.Fatalf("Error opening file/directory: %v", err)
-		}
-		defer func() {
-			err := windows.Close(handle)
-			if err != nil {
-				t.Logf("Error closing file %s: %v\n", testPath, err)
-			}
-		}()
-
-		extAttr, err := fs.GetFileEA(handle)
-		if err != nil {
-			t.Fatalf("Error getting extended attributes: %v", err)
-		}
-
-		test.Equals(t, len(testNode.ExtendedAttributes), len(extAttr), "Unexpected count of extended attribs for path: %s", testPath)
-
-		for _, expectedExtAttr := range testNode.ExtendedAttributes {
-			var foundExtAttr *fs.ExtendedAttribute
-			for _, ea := range extAttr {
-				if strings.EqualFold(ea.Name, expectedExtAttr.Name) {
-					foundExtAttr = &ea
-					break
-				}
-			}
-			test.Assert(t, foundExtAttr != nil, "Expected extended attribute not found")
-			test.Equals(t, expectedExtAttr.Value, foundExtAttr.Value, "Expected attribute value not equal for: %s", expectedExtAttr.Name)
-		}
-
-		sd, err := fs.GetFileSecurityDescriptor(testPath)
-
-		test.Assert(t, err == nil, "Error while getting the security descriptor")
-
-		testSD := string(testNode.GetGenericAttribute(restic.TypeSecurityDescriptor))
-		sdBytesTest, err := base64.StdEncoding.DecodeString(testSD)
-		if err != nil {
-			t.Fatalf("Error decoding SD: %s", err)
-		}
-		sdInput, err := fs.SecurityDescriptorBytesToStruct(sdBytesTest)
-
-		if err != nil {
-			t.Fatalf("Error converting SD to struct: %s", err)
-		}
-
-		sdBytesOutput, err := base64.StdEncoding.DecodeString(sd)
-		if err != nil {
-			t.Fatalf("Error decoding SD: %s", err)
-		}
-		sdOutput, err := fs.SecurityDescriptorBytesToStruct(sdBytesOutput)
-		if err != nil {
-			t.Fatalf("Error converting Output SD to struct: %s", err)
-		}
-
-		test.Equals(t, sdInput, sdOutput, "SecurityDescriptors not equal for path: %s", testPath)
-
-		fi, err := os.Lstat(testPath)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		nodeFromFileInfo, err := restic.NodeFromFileInfo(testPath, fi)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		sdNodeFromFileInfoInput := sdOutput
-
-		sdBytesFromNode := nodeFromFileInfo.GetGenericAttribute(restic.TypeSecurityDescriptor)
-
-		sdByteNodeOutput, err := base64.StdEncoding.DecodeString(string(sdBytesFromNode))
-		if err != nil {
-			t.Fatalf("Error decoding SD: %s", err)
-		}
-		sdNodeFromFileInfoOutput, err := fs.SecurityDescriptorBytesToStruct(sdByteNodeOutput)
-		if err != nil {
-			t.Fatalf("Error converting Output SD Through Node to struct: %s", err)
-		}
-
-		test.Equals(t, sdNodeFromFileInfoInput, sdNodeFromFileInfoOutput, "SecurityDescriptors got from NodeFromFileInfo not equal for path: %s", testPath)
-	}
-}
 
 func TestRestoreCreationTime(t *testing.T) {
 	tempDir := t.TempDir()
