+++ internal/restic/node_xattr.go
@@ -4,79 +4,29 @@
 package restic
 
 import (
-	"fmt"
 	"os"
 	"syscall"
 
-	"github.com/restic/restic/internal/debug"
 	"github.com/restic/restic/internal/errors"
 
 	"github.com/pkg/xattr"
 )
 
-func (node Node) restoreExtendedAttributes(path string) error {
-	for _, attr := range node.ExtendedAttributes {
-		err := setxattr(path, attr.Name, attr.Value)
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (node *Node) fillExtendedAttributes(path string) error {
-	xattrs, err := listxattr(path)
-	debug.Log("fillExtendedAttributes(%v) %v %v", path, xattrs, err)
-	if err != nil {
-		return err
-	}
-
-	node.ExtendedAttributes = make([]ExtendedAttribute, 0, len(xattrs))
-	for _, attr := range xattrs {
-		attrVal, err := getxattr(path, attr)
-		if err != nil {
-			fmt.Fprintf(os.Stderr, "can not obtain extended attribute %v for %v:\n", attr, path)
-			continue
-		}
-		attr := ExtendedAttribute{
-			Name:  attr,
-			Value: attrVal,
-		}
-
-		node.ExtendedAttributes = append(node.ExtendedAttributes, attr)
-	}
-
-	return nil
-}
-
-// restoreGenericAttributes is no-op.
-func (node *Node) restoreGenericAttributes(_ string) error {
-	for _, attr := range node.GenericAttributes {
-		handleUnknownGenericAttributeFound(attr.Name)
-	}
-	return nil
-}
-
-// fillGenericAttributes is a no-op.
-func (node *Node) fillGenericAttributes(_ string, _ os.FileInfo, _ *statT) error {
-	return nil
-}
-
-// getxattr retrieves extended attribute data associated with path.
-func getxattr(path, name string) ([]byte, error) {
+// Getxattr retrieves extended attribute data associated with path.
+func Getxattr(path, name string) ([]byte, error) {
 	b, err := xattr.LGet(path, name)
 	return b, handleXattrErr(err)
 }
 
-// listxattr retrieves a list of names of extended attributes associated with the
+// Listxattr retrieves a list of names of extended attributes associated with the
 // given path in the file system.
-func listxattr(path string) ([]string, error) {
+func Listxattr(path string) ([]string, error) {
 	l, err := xattr.LList(path)
 	return l, handleXattrErr(err)
 }
 
-// setxattr associates name and data together as an attribute of path.
-func setxattr(path, name string, data []byte) error {
+// Setxattr associates name and data together as an attribute of path.
+func Setxattr(path, name string, data []byte) error {
 	return handleXattrErr(xattr.LSet(path, name, data))
 }
 
@@ -97,4 +47,17 @@ func handleXattrErr(err error) error {
 	default:
 		return errors.WithStack(e)
 	}
+}
+
+// restoreGenericAttributes is no-op.
+func (node *Node) restoreGenericAttributes(_ string) error {
+	for _, attr := range node.GenericAttributes {
+		handleUnknownGenericAttributeFound(attr.Name)
+	}
+	return nil
+}
+
+// fillGenericAttributes is a no-op.
+func (node *Node) fillGenericAttributes(_ string, _ os.FileInfo, _ *statT) error {
+	return nil
 }
